INTERVENTION now provides direct attribute access to the known interventions.
To implement this, the lazy module now contains a second option for
adding lazy attributes to a class - necessary for dynamically defined
attributes.

Interventions can now handle additional logic to determine if an
intervention should be presented to a user (via the same
intervention.user_has_access() method).  Add additional access_strategy
methods to the intervention for such control.