Strategies for org association now use linked identifiers.
limit_by_clinic_list() and not_in_clinic_list() have been replaced with
versions that take identifer details.

This REQUIRES a change to site_persistence naming those methods.  All
orgs to be checked should be assigned an additional identifier which is
used in the check.  (A PR for TrueNTH site_persistence will follow and
must be in place when this goes live).