TN-977 correct off by one error (#2583)
* Correct off by one error.

Off by two, actually.  Zero index and strictly less than comparison.  Note, the failure is bumped AFTER the check if the current attempt is bad.

* Correct off by one error in test also.

* Given the need to allow a 5th attempt to work after 4 failures, return to strictly less than check.
This means on the 6th attempt, even a good password will fail and only on the 7th will the user see the dialog.
Logic needs to be reworked to implement as spec'd.

* Rework order of operations, so we can display the locked out message
when the 5th attempt fails.