IRONN-130-Spain-v5-all-orgs (#4279)
Switches these two Spanish orgs' v3 retired_as_of date to this coming Monday;
    PCCTC wants this interim state, as once we have the list of affected patients
    and timepoints for the change going farther back, they'll still need to
    discuss w/ the Spanish sites to see what's tolerable there. So this interim
    step allows for collection of future PROs according to v5.

Also removes a redundant org config for "Hospital Universitario Virgen de
la Victoria". The removed config is at v3, which is incorrect. The
remaining config has the valid transition to v5; this config was later
in the file, and it appears that it overrode the prior/bad config,
fortunately (at least, as I read
https://eproms.truenth.org/api/organization?tree_view=true today).

https://jira.movember.com/browse/IRONN-130

---------

Co-authored-by: Paul Bugni <pbugni@u.washington.edu>