TN-3097 confront random failures with 2FA tokens. (#4307)
Immediately attempt to use code before emailing out - regenerate if it
doesn't work.

I'll note, via unit testing I could NOT find a reliable way to get this
to break before or after changing, despite many attempts and huge
iteration counts. Therefore, I can't be certain this fixes the issue,
but I was able to verify by hand that some of the codes emailed out do
NOT work against the otp_secret saved. This has a good chance of being
the problem.