IRONN-223 Change a parent org to v5 (no impact) (#4352)
https://movember.atlassian.net/browse/IRONN-223
This org doesn't have any patients (and likely won't ever), it's used as
an umbrella for 3 sites that were transitioned to v5 3 years ago. This
change simply switches this parent to v5 with the same effective date as
the children, as a precaution in case staff do add patients here.