bug found in qb_timeline overlap check (#4368)
between protocol v3 and v5, additional months were added, i.e. month 33.

if v3 month 36 didn't have any committed work and it's time to change
protocols for the user, v5 month 33 should follow v3 month 30.

the code looking for this case had an extra check which prevented the
above from functioning, in the rare case where v3 month 30 ended before
v5 month 33 starts.

this was discovered with the new withdrawal handling in adherence data,
as the insertion of the withdrawal row could not predictably insert
before the time point just after withdrawal, when the timeline was found
to have such an overlap.