TN-3353 apptext caching doesn't consider locale.  disable apptext cache. (#4455)
As TN-3353 demonstrates, it's possible for a cached apptext value to be
translated using a different locale from the current user's preferences.
Given the complexity of altering the calling parameters to look up
`app_text` to include locale, and make the cache i18n proof - and the
rare use of apptext - opting to disable the cache so the current user's
locale should always be consulted.